requirements.txt
opencv-python==4.11.0.86
moviepy==2.1.2
transformers==4.50.3
pillow==10.4.0
torch==2.6.0
scipy==1.15.2
requests==2.32.3
colorama==0.4.6